refactor(tools): migrate pipeline to the agent_tools registry (#3629)
